Change match aggregation to a class, move it and protein ID mapper to separate folders
