+++ sponge/modules/match_aggregator/__init__.py
@@ -1,0 +1,1 @@
+from sponge.modules.match_aggregator.match_aggregator import MatchAggregator
