+++ sponge/modules/motif_selector/__init__.py
@@ -1,2 +1,1 @@
 from sponge.modules.motif_selector.motif_selector import MotifSelector
-from sponge.modules.motif_selector.protein_id_mapper import ProteinIDMapper
