+++ sponge/modules/utils/__init__.py
@@ -1,5 +1,4 @@
 from .data_retrieval import *
-from .filtering_functions import *
-from .match_aggregation import *
+from .tfbs_filtering import *
 from .motif_information import *
 from .string_manipulation import *
